Move `proc_params()` to separate file
