+++ benchmarker/modules/problems/cosmoflow/tensorflow.py
@@ -7,11 +7,10 @@ defined in cosmoflow.py.
 
 """
 
-import argparse
-import ast
-
 import tensorflow as tf
 import tensorflow.keras.layers as layers
+
+from .params import proc_params
 
 
 def scale_1p2(x):
@@ -21,18 +20,6 @@ def scale_1p2(x):
     tanh activation for targets in the range [-1,1].
     """
     return x * 1.2
-
-
-def proc_params(params, unparsed_args):
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--input_shape", default="128, 128, 128, 4")
-    parser.add_argument("--target_size", default=4)
-    parser.add_argument("--dropout", default=0)
-    args, unparsed = parser.parse_known_args(unparsed_args)
-    params["input_shape"] = ast.literal_eval(args.input_shape)
-    params["target_size"] = args.target_size
-    params["dropout"] = args.dropout
-    assert unparsed == []
 
 
 def build_model(input_shape, target_size, dropout=0):
